feat: video card added
